+++ qadence/__init__.py
@@ -5,6 +5,7 @@ from importlib import import_module
 from torch import cdouble, set_default_dtype
 from torch import float64 as torchfloat64
 
+from .analog import *
 from .backend import *
 from .backends import *
 from .blocks import *
@@ -28,6 +29,7 @@ from .utils import *
 DEFAULT_FLOAT_DTYPE = torchfloat64
 DEFAULT_COMPLEX_DTYPE = cdouble
 set_default_dtype(DEFAULT_FLOAT_DTYPE)
+
 """
 The imports above fetch the functions defined in the __all__ of each sub-module
 to the qadence name space. Make sure each added submodule has the respective definition:
@@ -39,6 +41,7 @@ the __all__ of the qadence namespace. Make sure to keep alphabetical ordering.
 """
 
 list_of_submodules = [
+    ".analog",
     ".backends",
     ".blocks",
     ".circuit",
