+++ qadence/analog/__init__.py
@@ -1,0 +1,5 @@
+from __future__ import annotations
+
+from .emulate import add_interaction
+
+__all__ = ["add_interaction"]
