+++ qadence/transpile/__init__.py
@@ -10,8 +10,7 @@ from .block import (
 )
 from .circuit import fill_identities
 from .digitalize import digitalize
-from .emulate import add_interaction
 from .invert import invert_endianness, reassign
 from .transpile import blockfn_to_circfn, transpile
 
-__all__ = ["add_interaction", "set_trainable", "invert_endianness"]
+__all__ = ["set_trainable", "invert_endianness"]
